Weekly commit 25/09/22
This week I worked on tests, as well as added a /dev/ subrepo. It is mostly a copy of the current repository, but with heavy modifications to the hierarchical model's input. Essentially, I am trying to make a sentence-level hierarchical model. Once I am done I will integrate this to the regular scripts, under a new name or new options.